+++ src/engine/idTable/IdTable.h
@@ -13,6 +13,7 @@
 #include <vector>
 
 #include "engine/idTable/IdTableRow.h"
+#include "engine/idTable/ResizeWhenMoveVector.h"
 #include "global/Id.h"
 #include "util/Algorithm.h"
 #include "util/AllocatorWithLimit.h"
@@ -102,48 +103,6 @@ namespace columnBasedIdTable {
 // TODO<joka921> The NumColumns should be `size_t` but that requires several
 // additional changes in the rest of the code.
 //
-
-namespace detail {
-// Disclaimer: This class is an implementation detail of the column based ID
-// tables. Its semantics are very particular, so we don't expect it to have a
-// use case outside of the `IdTable` module.
-// A class that inherits from a `vector<T>`, where `T`.
-// This class changes the move operators of the underlying
-// `vector` as follows: Instead of moving the vector as a whole, only the
-// indidual elements are moved. This is used for the column based `IdTables`
-// where we move the indivdual columns, but still want a moved from table to
-// have the same number of columns as before, but with the columns now being
-// empty.
-template <typename T>
-struct ResizeWhenMoveVector : public std::vector<T> {
-  using Base = std::vector<T>;
-  using Base::Base;
-  // Defaulted copy operations
-  ResizeWhenMoveVector(const ResizeWhenMoveVector&) = default;
-  ResizeWhenMoveVector& operator=(const ResizeWhenMoveVector&) = default;
-
-  // Move operations with the specified semantics.
-  ResizeWhenMoveVector(ResizeWhenMoveVector&& other) noexcept {
-    moveImpl(std::move(other));
-  }
-  ResizeWhenMoveVector& operator=(ResizeWhenMoveVector&& other) noexcept {
-    this->clear();
-    moveImpl(std::move(other));
-    return *this;
-  }
-
- private:
-  // The common implementation, move the elements of `other` into `this`. Terminate with a readable error message in the unlikely case of `std::bad_alloc`.
-  void moveImpl(ResizeWhenMoveVector&& other) noexcept {
-    ad_utility::terminateIfThrows(
-        [&other, self = this] {
-          self->insert(self->end(), std::make_move_iterator(other.begin()),
-                       std::make_move_iterator(other.end()));
-        },
-        "Error happened during the move construction or move assignment of an IdTable");
-  }
-};
-}
 template <typename T = Id, int NumColumns = 0,
           typename ColumnStorage = std::vector<
               T, ad_utility::default_init_allocator<T, std::allocator<T>>>,
